chore(MoneyKit): IOS-4760 Cleanup and refactor - Part 8 (#2641)
* IOS-4760: Migrate PriceService to Combine

* IOS-4760: Rename priceInFiat argument label to exchangeRate

* IOS-4760: Rename MoneyValuePair init arguments

* IOS-4760: Move window argument to last position in priceSeries and priceHistoric

* IOS-4760: Rename price argument label from 'for' to 'of'

* IOS-4760: Refactor and document Currency

* IOS-4760: Refactor and document Money

* IOS-4760: Refactor and document Crypto and Fiat

* IOS-4760: Refactor and document Price

* IOS-4760: Refactor and document HistoricalFiatPriceService

* IOS-4760: Use non-failing MoneyImplementing.create where possible

* IOS-4760: Pass fiatCurrency in FiatFormatterProvider

* IOS-4760: Pass amount to CryptoFormatter.format

* IOS-4760: Set CryptoFormatter precision on init

* IOS-4760: Remove redundant self wrapper

* IOS-4760: Nits

* IOS-4760: Fix inconsistencies between Currency and CurrencyType

* IOS-4760: Fix precision loss due to displayMajorValue

* IOS-4760: Replace PriceQuoteAtTimeResponse usages with PriceQuoteAtTime

* IOS-4760: Nits

* IOS-4760: PR comments

* IOS-4760: Inline convertToFiatValue and convertToCryptoValue

* IOS-4760: Use displayString where appropaite

* IOS-4760: Nit

* IOS-4760-8: Fix after merge

* IOS-4760-8: Fix after merge

* IOS-4760-8: Fix after merge

* IOS-4760-8: Fix after merge

* IOS-4760-8: Fix after merge

* IOS-4760-8: Fix after merge

* IOS-4760: 8: Nit

* IOS-4760: Fix build errors

* IOS-4760: Fix exchange rate calculation

* Update Modules/Tool/Sources/ToolKit/General/Extensions/Decimal+Extensions.swift

* Fix tests

+++ Modules/FeatureDashboard/Sources/FeatureDashboardUI/Components/TotalBalanceTableViewCell/TotalBalanceViewPresenter.swift
@@ -64,7 +64,7 @@ extension PortfolioBalanceChangeProvider: AssetPriceViewInteracting {
                         next: .init(
                             currentPrice: change.balance,
                             time: .hours(24),
-                            changePercentage: change.changePercentage,
+                            changePercentage: change.changePercentage.doubleValue,
                             priceChange: change.change
                         )
                     )
